AC-18 #comment Added HeartButton as composable
