+++ presentation/ui/src/main/java/com/jfmr/ac/test/presentation/ui/character/detail/view/CharacterDetailScreen.kt
@@ -1,9 +1,5 @@
 package com.jfmr.ac.test.presentation.ui.character.detail.view
 
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.IntrinsicSize
@@ -12,14 +8,11 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.FavoriteBorder
-import androidx.compose.material.icons.outlined.Favorite
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -29,17 +22,14 @@ import androidx.compose.material3.SmallTopAppBar
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.scale
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import coil.compose.AsyncImage
@@ -51,7 +41,7 @@ import com.jfmr.ac.test.presentation.ui.character.detail.viewmodel.DetailViewMod
 import com.jfmr.ac.test.presentation.ui.episode.list.view.EpisodesScreen
 import com.jfmr.ac.test.presentation.ui.main.component.CircularProgressBar
 import com.jfmr.ac.test.presentation.ui.main.component.ErrorScreen
-import kotlinx.coroutines.launch
+import com.jfmr.ac.test.presentation.ui.main.component.HeartButton
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -94,13 +84,6 @@ private fun CharacterDetailContent(
     var scrolledY = 0f
     var previousOffset = 0
 
-    val interactionSource = MutableInteractionSource()
-
-    val coroutineScope = rememberCoroutineScope()
-    val scale = remember {
-        Animatable(1f)
-    }
-
     Box(modifier = Modifier.fillMaxWidth(), contentAlignment = Alignment.Center) {
         LazyColumn(modifier = Modifier.fillMaxWidth(), verticalArrangement = Arrangement.SpaceBetween) {
             item {
@@ -119,7 +102,9 @@ private fun CharacterDetailContent(
                             previousOffset = lazyListState.firstVisibleItemScrollOffset
                         }
                         .clip(CircleShape),
-                    contentScale = ContentScale.FillWidth)
+                    contentScale = ContentScale.FillWidth,
+                    error = painterResource(id = R.drawable.ic_placeholder)
+                )
             }
             item {
                 Row(
@@ -134,33 +119,7 @@ private fun CharacterDetailContent(
                             text = stringResource(id = R.string.character),
                             style = MaterialTheme.typography.titleLarge,
                         )
-                        Icon(
-                            imageVector = if (character.isFavorite == true) {
-                                Icons.Outlined.Favorite
-                            } else {
-                                Icons.Default.FavoriteBorder
-                            },
-                            contentDescription = stringResource(id = R.string.fav_description),
-                            tint = Color.Red,
-                            modifier = Modifier
-                                .scale(scale = scale.value)
-                                .size(size = dimensionResource(id = R.dimen.favorite_size))
-                                .align(Alignment.TopEnd)
-                                .padding(dimensionResource(id = R.dimen.character_list_padding))
-                                .clickable(interactionSource = interactionSource, indication = null) {
-                                    coroutineScope.launch {
-                                        scale.animateTo(
-                                            0.8f,
-                                            animationSpec = tween(100),
-                                        )
-                                        scale.animateTo(
-                                            1f,
-                                            animationSpec = tween(100),
-                                        )
-                                    }
-                                    action(character.copy(isFavorite = !character.isFavorite!!))
-                                }
-                        )
+                        HeartButton(character, action, Alignment.TopEnd)
                     }
                 }
             }
@@ -192,6 +151,7 @@ private fun CharacterDetailContent(
         }
     }
 }
+
 
 @Composable
 private fun DetailRow(nameResource: Int, value: String) {
